ref: Refactor react-router-based adapters into a shared implementation
